build: lint test code, and fail the gate on warnings
`make clippy` compiled `--bin chm` and the hypervisor lib, so every
`#[cfg(test)]` module was invisible to it: 842 test functions in chm and
326 in hypervisor, unlinted. `arch` was the only crate passing
`--all-targets`, and it has 32.

That is how an orphaned `#[test]` reached main three times. rustc
diagnosed it on each occasion, as `duplicated attribute` and then
`dead_code` on the helper left stranded behind it. The diagnosis went
into code the gate never built.

Adding `--all-targets` alone was not enough, and this is the part worth
recording: both of those diagnostics are warn-level, so the gate printed
them and still exited 0. Measured, on a duplicated `#[test]`: exit 0
without `-D warnings`, exit 101 with it. A gate that reports a defect and
passes anyway is a log line.

The backlog `--all-targets` exposed was 367 lints and contained no dead
guards and no bugs. 282 were `absolute_paths` inside test bodies, and a
further 8 were `identity_op` on deliberate register arithmetic --
`itt + 0 * 8` next to `itt + 1 * 8` to show the ITT stride, `(intid <<
16) | 0` to show a field that is architecturally zero. Folding those away
would have hidden the layout the tests exist to pin down, so they are
allowed under cfg(test) with the reason written down rather than
"fixed". Shipped code keeps every denial: `--all-targets` also builds the
non-test target, where cfg(test) is off.

Two of the assertions clippy flagged as constant-valued were checked
rather than assumed. Both still panic when the constant they guard moves,
so both guards work and neither was disabled.

Real defects fixed on the way through: an unsafe block missing its
SAFETY comment, a doc comment attached to nothing, two
`Default::default()` field reassignments, an `.err().expect()`, a
`let _ =` on a unit return, and two needless range loops. The unreachable
match arm in hvf_boot.rs was kept and annotated, not deleted: it is
unreachable only because `CpuState`'s variants are feature-gated, and it
is what keeps the match exhaustive once `kvm` or `mshv` is enabled.

Mutation-proved, exit codes from the full three-line recipe:

  mutation                                    old gate   new gate
  orphaned test helper (dead_code)                   0          2
  duplicated #[test] (the shipped bug)               0          2
  unused_mut in tests/hvf_boot.rs                    0          2
  absolute_paths in production code                  2          2

The last row is the control: production linting is unchanged by the
cfg(test) allowances.

836 chm tests and 263 hypervisor lib tests pass; hvf_boot still builds.

+++ chm/src/vanilla_export.rs
@@ -649,7 +649,7 @@ mod tests {
             patch_clock(&mut doc, &st, &ancestor).expect("the clock must be writable");
 
             let got = doc.clock().unwrap();
-            let want = base.cntvct + secs * u64::from(base.cntfrq);
+            let want = base.cntvct + secs * base.cntfrq;
             assert_eq!(got.cntvct, want, "after {secs}s the counter must advance by secs*cntfrq");
             assert_eq!(got.host_realtime_ns, now, "the capture instant must be the checkpoint's");
             assert_eq!(
